Move the gateway spec helper into support folder
